requirements.txt
aiofiles==23.2.1
aiolimiter==1.1.0
anyio==4.3.0
APScheduler==3.10.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
nats-py==2.7.2
packaging==24.0
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
tornado==6.4
tzlocal==5.2
